refactor: client code
- Added new DashboardContent component to manage page listings and interactions.
- Introduced PageCard component for displaying individual pages with delete functionality.
- Implemented useNowTick hook for real-time updates in the dashboard.
- Created various components for the upload form, including Features, LoginModal, PreviewDialog, and SuccessView.
- Integrated TurnstileWidget for enhanced security during uploads.
- Updated routing to reflect new component structure and improve user experience.

This update significantly enhances the functionality and usability of the dashboard and upload features.
