+++ client/features/upload-form/components/PreviewDialog.tsx
@@ -1,8 +1,8 @@
 import { Close } from "@hugeicons/core-free-icons";
 import { HugeiconsIcon } from "@hugeicons/react";
 import { useCallback } from "react";
+import { cn } from "@/utils/styles";
 import { usePreview } from "../hooks/usePreview";
-import { cn } from "../utils/styles";
 
 interface PreviewDialogProps {
 	file: File;
