+++ client/lib/i18nResources.ts
@@ -70,7 +70,6 @@ export const resources = {
 				},
 				noDescription: "No description",
 				delete: "Delete",
-				deleting: "Deleting...",
 				loadingPages: "Loading pages…",
 				noExpiredPagesFound: "No expired pages found",
 				noPagesYet: "No pages yet",
@@ -155,7 +154,6 @@ export const resources = {
 				},
 				noDescription: "설명이 없습니다",
 				delete: "삭제",
-				deleting: "삭제 중…",
 				loadingPages: "페이지 불러오는 중…",
 				noExpiredPagesFound: "만료된 페이지가 없습니다",
 				noPagesYet: "아직 페이지가 없습니다",
@@ -238,7 +236,6 @@ export const resources = {
 				},
 				noDescription: "暂无描述",
 				delete: "删除",
-				deleting: "正在删除…",
 				loadingPages: "正在加载页面…",
 				noExpiredPagesFound: "未找到已过期页面",
 				noPagesYet: "还没有页面",
