+++ client/tsconfig.json
@@ -23,6 +23,7 @@
 		"noUncheckedSideEffectImports": true,
 
 		"paths": {
+			"@/*": ["./*"],
 			"@shared/*": ["../shared/*"]
 		}
 	},
