+++ server/tsconfig.json
@@ -40,6 +40,7 @@
 		"types": ["../worker-configuration.d.ts"],
 
 		"paths": {
+			"@/*": ["./*"],
 			"@shared/*": ["../shared/*"]
 		}
 	},
