+++ vite.config.ts
@@ -18,6 +18,7 @@ export default defineConfig({
 	],
 	resolve: {
 		alias: {
+			"@": resolve(__dirname, "./client"),
 			"@shared": resolve(__dirname, "./shared"),
 		},
 	},
